Feat: 에러 바운더리 완료 및 테스트 완료
